fix bugs related to navigation and user permissions
