--- src/layouts/CleanLayout.jsx
@@ -1,20 +1,0 @@
-import Sidebar from '../shared/components/Sidebar'
-import VerifyUser from '../pages/VerifyUser'
-
-const CleanLayout = ({ children }) => {
-  return (
-    <div className="flex min-h-screen flex-col">
-  <header className=" p-4"><VerifyUser /></header>
-  <div className="flex flex-1 flex-row">
-    <aside className="w-1/3 p-4 max-w-[18rem]">
-      <Sidebar
-      />
-    </aside>
-    {children}
-  </div>
-</div>
-
-  )
-}
-
-export default CleanLayout
