+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import PostList from '../features/posts/PostList'
-import HolyGrailLayout from '../layouts/HolyGrailLayout'
+import HolyGrailLayout from '../layouts/SearchLayout'
 import { ProfileOption } from '../shared/components/ProfileOption'
 
 const Home = () => {
