+++ src/pages/PublicUserProfile.jsx
@@ -1,4 +1,4 @@
-import CleanLayout from '../layouts/CleanLayout'
+import CleanLayout from '../layouts/HolyGrailLayout'
 import PublicProfileComponent from '../features/publicProfile/publicProfileComponent'
 const PublicUserProfile = () => {
   return (
